+++ .changeset/funny-news-pay.md
@@ -1,0 +1,6 @@
+---
+'@syncit/transporter': patch
+'@syncit/core': patch
+---
+
+Move Transporter interface to @syncit/core to remove circular dependencies
