+++ packages/core/package.json
@@ -56,7 +56,6 @@
   },
   "dependencies": {
     "@rrweb/types": "^2.0.0-alpha.5",
-    "@syncit/transporter": "^1.0.0",
     "xstate": "^4.37.0",
     "rrweb": "^2.0.0-alpha.5",
     "rrweb-snapshot": "^2.0.0-alpha.5"
