+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
 export * from './buffer';
 export * from './common';
 export * from './machine';
+export * from './transporter';
