+++ packages/core/src/machine.ts
@@ -3,7 +3,7 @@
 import { createMachine, interpret, assign, EventObject } from 'xstate';
 import { Replayer, record } from 'rrweb';
 import { listenerHandler, eventWithTime } from '@rrweb/types';
-import { Transporter } from '@syncit/transporter';
+import { Transporter } from './transporter';
 import { SourceBuffer, Chunk } from './buffer';
 import { onMirror, RemoteControlActions, CustomEventTags } from './common';
